requirements.txt
fastapi
psycopg2
requests
pydantic
google-generativeai
uvicorn
